🚚🚚🚚🚚🚚 Move all installers to separate files
