+++ src/installers.ts
@@ -1,21 +1,11 @@
-import fs from "fs";
 import { win32 } from "path";
-import KeyTree from "key-tree";
 import {
-  findAllSubdirsWithSome,
   FileTree,
   fileTreeFromPaths,
   FILETREE_ROOT,
-  filesIn,
-  filesUnder,
-  pathInTree,
-  findDirectSubdirsWithSome,
-  dirWithSomeIn,
-  fileCount,
   sourcePaths,
   subdirNamesIn,
   subtreeFrom,
-  Glob,
 } from "./filetree";
 import {
   VortexApi,
@@ -30,45 +20,7 @@ import {
   VortexInstallFunc,
   VortexInstruction,
 } from "./vortex-wrapper";
-import {
-  CET_GLOBAL_INI,
-  CET_MOD_CANONICAL_INIT_FILE,
-  CET_MOD_CANONICAL_PATH_PREFIX,
-  RED4EXT_CORE_RED4EXT_DLL,
-  RED4EXT_MOD_CANONICAL_BASEDIR,
-  RED4EXT_KNOWN_NONOVERRIDABLE_DLLS,
-  RED4EXT_KNOWN_NONOVERRIDABLE_DLL_DIRS,
-  Red4ExtLayout,
-  REDS_MOD_CANONICAL_EXTENSION,
-  REDS_MOD_CANONICAL_PATH_PREFIX,
-  RESHADE_MOD_PATH,
-  RESHADE_SHADERS_DIR,
-  INI_MOD_EXT,
-  INI_MOD_PATH,
-  JSON_FILE_EXT,
-  KNOWN_JSON_FILES,
-  ASI_MOD_EXT,
-  ASI_MOD_PATH,
-  LayoutToInstructions,
-  NoInstructions,
-  MaybeInstructions,
-  InvalidLayout,
-  CetLayout,
-  RedscriptLayout,
-  AsiLayout,
-  KNOWN_TOPLEVEL_DIRS,
-} from "./installers.layouts";
-import {
-  toSamePath,
-  toDirInPath,
-  instructionsForSourceToDestPairs,
-  instructionsForSameSourceAndDestPaths,
-  useFirstMatchingLayoutForInstructions,
-  moveFromTo,
-  useAllMatchingLayouts,
-  makeSyntheticName,
-} from "./installers.shared";
-import { showRed4ExtReservedDllErrorDialog } from "./dialogs";
+import { KNOWN_TOPLEVEL_DIRS } from "./installers.layouts";
 import {
   testForCetCore,
   installCetCore,
@@ -81,24 +33,20 @@ import {
   testCoreWolvenKitCli,
   installCoreWolvenkit,
 } from "./installer.core";
-import { trueish } from "./installers.utils";
 import { GAME_ID } from "./index.metadata";
 import { Installer, InstallerType, InstallerWithPriority } from "./installers.types";
 import { installCoreTweakXL, testForCoreTweakXL } from "./installer.core-tweak-xl";
-import {
-  promptToFallbackOrFailOnUnresolvableLayout,
-  testForFallback,
-  installFallback,
-} from "./installer.fallback";
+import { testForFallback, installFallback } from "./installer.fallback";
 import { installTweakXLMod, testForTweakXLMod } from "./installer.tweak-xl";
 import { installCoreArchiveXL, testForCoreArchiveXL } from "./installer.core-archive-xl";
-import {
-  extraCanonArchiveInstructions,
-  testForArchiveMod,
-  installArchiveMod,
-  archiveCanonLayout,
-  archiveHeritageLayout,
-} from "./installer.archive";
+import { testForArchiveMod, installArchiveMod } from "./installer.archive";
+import { testForMultiTypeMod, installMultiTypeMod } from "./installer.multitype";
+import { testForAsiMod, installAsiMod } from "./installer.asi";
+import { testForCetMod, installCetMod } from "./installer.cet";
+import { testForIniMod, installIniMod } from "./installer.config.ini-reshade";
+import { testForJsonMod, installJsonMod } from "./installer.config.json";
+import { testForRed4ExtMod, installRed4ExtMod } from "./installer.red4ext";
+import { testForRedscriptMod, installRedscriptMod } from "./installer.redscript";
 
 // Ensure we're using win32 conventions
 const path = win32;
@@ -128,1008 +76,7 @@ export const notInstallableMod: VortexWrappedInstallFunc = (
   throw new Error("Should never get here");
 };
 
-const matchAsiFile = (file: string) => path.extname(file) === ASI_MOD_EXT;
-
-const matchRedscript = (file: string) =>
-  path.extname(file) === REDS_MOD_CANONICAL_EXTENSION;
-
-const allRedscriptFiles = (files: string[]): string[] => files.filter(matchRedscript);
-
 // Installers
-const allFilesInFolder = (folder: string, files: string[]) => {
-  const fileTree = new KeyTree({ separator: path.sep });
-
-  files.forEach((file) => fileTree.add(file, file));
-
-  const moddir = fileTree._getNode(folder); // eslint-disable-line no-underscore-dangle
-
-  if (!moddir) {
-    return [];
-  }
-
-  const moddirPath = path.join(...moddir.fullPath);
-
-  const allTheFiles: string[] = [].concat(
-    ...Object.values(fileTree.getSub(moddirPath, true)),
-  );
-
-  return allTheFiles;
-};
-
-const allCanonicalCetFiles = (files: string[]) =>
-  allFilesInFolder(CET_MOD_CANONICAL_PATH_PREFIX, files);
-
-//
-//
-//
-//
-// Installers
-//
-// These should come in (roughly) reverse order of priority,
-// because the highest-priority ones will use Layouts and
-// other parts from the simpler installers.
-//
-//
-
-// CET
-
-const matchCetInitLua = (f: string): boolean =>
-  path.basename(f) === CET_MOD_CANONICAL_INIT_FILE;
-
-const findCanonicalCetDirs = (fileTree: FileTree): string[] =>
-  findDirectSubdirsWithSome(CET_MOD_CANONICAL_PATH_PREFIX, matchCetInitLua, fileTree);
-
-const detectCetCanonLayout = (fileTree: FileTree): boolean =>
-  findCanonicalCetDirs(fileTree).length > 0;
-
-const cetCanonLayout = (
-  api: VortexApi,
-  _modName: string,
-  fileTree: FileTree,
-): MaybeInstructions => {
-  const allCanonCetFiles = findCanonicalCetDirs(fileTree).flatMap((namedSubdir) =>
-    filesUnder(namedSubdir, Glob.Any, fileTree),
-  );
-
-  if (allCanonCetFiles.length < 1) {
-    api.log("debug", "No canonical CET files found.");
-    return NoInstructions.NoMatch;
-  }
-
-  return {
-    kind: CetLayout.Canon,
-    instructions: instructionsForSameSourceAndDestPaths(allCanonCetFiles),
-  };
-};
-
-// CET
-
-// CET mods are detected by:
-//
-// Canonical:
-//  - .\bin\x64\plugins\cyber_engine_tweaks\mods\MODNAME\init.lua
-//  - .\r6\scripts\[modname]\*.reds
-//
-// Fixable: no
-//
-// Archives: both canonical
-
-export const testForCetMod: VortexWrappedTestSupportedFunc = (
-  _api: VortexApi,
-  log: VortexLogFunc,
-  _files: string[],
-  fileTree: FileTree,
-): Promise<VortexTestResult> => {
-  const hasCetFilesInANamedModDir = detectCetCanonLayout(fileTree);
-
-  if (!hasCetFilesInANamedModDir) {
-    return Promise.resolve({ supported: false, requiredFiles: [] });
-  }
-
-  log("info", `Matching CET installer: ${hasCetFilesInANamedModDir}`);
-
-  return Promise.resolve({
-    supported: hasCetFilesInANamedModDir,
-    requiredFiles: [],
-  });
-};
-
-// Install the CET stuff, as well as any archives we find
-export const installCetMod: VortexWrappedInstallFunc = (
-  api: VortexApi,
-  log: VortexLogFunc,
-  files: string[],
-  fileTree: FileTree,
-  _destinationPath: string,
-): Promise<VortexInstallResult> => {
-  const cetFiles = allCanonicalCetFiles(files);
-
-  if (cetFiles.length === 0) {
-    return Promise.reject(
-      new Error("CET install but no CET files, should never get here"),
-    );
-  }
-
-  const extraArchiveInstructions = extraCanonArchiveInstructions(api, fileTree);
-
-  const instructions = [
-    ...instructionsForSameSourceAndDestPaths(cetFiles),
-    ...extraArchiveInstructions.instructions,
-  ];
-
-  return Promise.resolve({ instructions });
-};
-
-// Reds
-
-// Redscript mods are detected by
-//
-// Canonical:
-//  - .\r6\scripts\[modname]\*.reds
-//
-// Fixable (everything assumed to be one mod):
-//  - .\*.reds
-//  - .\r6\scripts\*.reds
-//
-// Archives:
-//  - Canonical both only
-
-const detectRedscriptBasedirLayout = (fileTree: FileTree): boolean =>
-  dirWithSomeIn(REDS_MOD_CANONICAL_PATH_PREFIX, matchRedscript, fileTree);
-
-const redscriptBasedirLayout = (
-  api: VortexApi,
-  modName: string,
-  fileTree: FileTree,
-): MaybeInstructions => {
-  const hasBasedirReds = detectRedscriptBasedirLayout(fileTree);
-
-  if (!hasBasedirReds) {
-    api.log("debug", "No basedir Redscript files found");
-    return NoInstructions.NoMatch;
-  }
-
-  const allBasedirAndSubdirFiles = filesUnder(
-    REDS_MOD_CANONICAL_PATH_PREFIX,
-    Glob.Any,
-    fileTree,
-  );
-
-  const modnamedDir = path.join(REDS_MOD_CANONICAL_PATH_PREFIX, modName);
-
-  const allToBasedirWithSubdirAsModname = allBasedirAndSubdirFiles.map(
-    moveFromTo(REDS_MOD_CANONICAL_PATH_PREFIX, modnamedDir),
-  );
-
-  return {
-    kind: RedscriptLayout.Basedir,
-    instructions: instructionsForSourceToDestPairs(allToBasedirWithSubdirAsModname),
-  };
-};
-
-const findCanonicalRedscriptDirs = (fileTree: FileTree) =>
-  findDirectSubdirsWithSome(REDS_MOD_CANONICAL_PATH_PREFIX, matchRedscript, fileTree);
-
-const detectRedscriptCanonOnlyLayout = (fileTree: FileTree): boolean =>
-  !detectRedscriptBasedirLayout(fileTree) &&
-  findCanonicalRedscriptDirs(fileTree).length > 0;
-
-const redscriptCanonLayout = (
-  api: VortexApi,
-  _modName: string,
-  fileTree: FileTree,
-): MaybeInstructions => {
-  const allCanonRedscriptFiles = findCanonicalRedscriptDirs(fileTree).flatMap(
-    (namedSubdir) => filesUnder(namedSubdir, Glob.Any, fileTree),
-  );
-
-  if (allCanonRedscriptFiles.length < 1) {
-    api.log("error", "No canonical Redscript files found.");
-    return NoInstructions.NoMatch;
-  }
-
-  // This is maybe slightly annoying to check, but makes
-  // logic elsewhere cleaner. I suppose we can decide that
-  // layouts need to be robust enough in themselves if they
-  // would otherwise depend on some external check that isn't
-  // always present.
-  //
-  // Generally, shouldn't get here.
-  //
-  const hasBasedirReds = detectRedscriptBasedirLayout(fileTree);
-
-  if (hasBasedirReds) {
-    // Errors need to be handled downstream if it's relevant there
-    api.log("debug", "No instructions from canon: basedir overrides");
-    return NoInstructions.NoMatch;
-  }
-
-  return {
-    kind: RedscriptLayout.Canon,
-    instructions: instructionsForSameSourceAndDestPaths(allCanonRedscriptFiles),
-  };
-};
-
-export const testForRedscriptMod: VortexWrappedTestSupportedFunc = (
-  api: VortexApi,
-  log: VortexLogFunc,
-  files: string[],
-  _fileTree: FileTree,
-): Promise<VortexTestResult> => {
-  const redscriptFiles = allRedscriptFiles(files);
-
-  log("debug", "redscriptFiles: ", { redscriptFiles });
-
-  // We could do more detection here but the
-  // installer will already need to duplicate
-  // all that. Maybe just check whether there
-  // are any counterindications?
-  if (redscriptFiles.length === 0) {
-    log("debug", "No Redscripts");
-    return Promise.resolve({ supported: false, requiredFiles: [] });
-  }
-
-  log("info", "Matched to Redscript");
-  return Promise.resolve({
-    supported: true,
-    requiredFiles: [],
-  });
-};
-
-// Install the Redscript stuff, as well as any archives we find
-export const installRedscriptMod: VortexWrappedInstallFunc = async (
-  api: VortexApi,
-  log: VortexLogFunc,
-  files: string[],
-  fileTree: FileTree,
-  destinationPath: string,
-): Promise<VortexInstallResult> => {
-  // We could get a lot fancier here, but for now we don't accept
-  // subdirectories anywhere other than in a canonical location.
-
-  // .\*.reds
-  // eslint-disable-next-line no-underscore-dangle
-  const hasToplevelReds = dirWithSomeIn(FILETREE_ROOT, matchRedscript, fileTree);
-  const toplevelReds = hasToplevelReds
-    ? filesUnder(FILETREE_ROOT, Glob.Any, fileTree)
-    : [];
-
-  // .\r6\scripts\*.reds
-  // eslint-disable-next-line no-underscore-dangle
-  const hasBasedirReds = dirWithSomeIn(
-    REDS_MOD_CANONICAL_PATH_PREFIX,
-    matchRedscript,
-    fileTree,
-  );
-  const basedirReds = hasBasedirReds
-    ? filesUnder(REDS_MOD_CANONICAL_PATH_PREFIX, Glob.Any, fileTree)
-    : [];
-
-  const canonSubdirs = findDirectSubdirsWithSome(
-    REDS_MOD_CANONICAL_PATH_PREFIX,
-    matchRedscript,
-    fileTree,
-  );
-  const hasCanonReds = canonSubdirs.length > 0;
-  const canonReds = hasCanonReds
-    ? canonSubdirs.flatMap((dir) => filesUnder(dir, Glob.Any, fileTree))
-    : [];
-
-  const installable = [hasToplevelReds, hasBasedirReds, hasCanonReds].filter(trueish);
-
-  if (installable.length !== 1) {
-    return promptToFallbackOrFailOnUnresolvableLayout(
-      api,
-      InstallerType.Redscript,
-      fileTree,
-    );
-  }
-
-  const modName = makeSyntheticName(destinationPath);
-
-  const extraArchiveInstructions = extraCanonArchiveInstructions(api, fileTree);
-
-  // Only one of these should exist but why discriminate?
-  const allSourcesAndDestinations = [
-    canonReds.map(toSamePath),
-    basedirReds.map(toDirInPath(REDS_MOD_CANONICAL_PATH_PREFIX, modName)),
-    toplevelReds.map(toDirInPath(REDS_MOD_CANONICAL_PATH_PREFIX, modName)),
-  ];
-
-  const redsInstructions = allSourcesAndDestinations.flatMap(
-    instructionsForSourceToDestPairs,
-  );
-
-  const instructions = [...redsInstructions, ...extraArchiveInstructions.instructions];
-
-  return Promise.resolve({ instructions });
-};
-
-// Red4Ext
-
-const matchDll = (file: string) => path.extname(file) === ".dll";
-const reservedDllDir = (dir: string) =>
-  RED4EXT_KNOWN_NONOVERRIDABLE_DLL_DIRS.includes(path.join(dir, path.sep));
-const reservedDllName = (file: string) =>
-  RED4EXT_KNOWN_NONOVERRIDABLE_DLLS.includes(path.join(file));
-
-const findBasedirRed4ExtFiles = (fileTree: FileTree) =>
-  filesUnder(RED4EXT_MOD_CANONICAL_BASEDIR, Glob.Any, fileTree);
-
-const detectRed4ExtBasedirLayout = (fileTree: FileTree): boolean =>
-  dirWithSomeIn(RED4EXT_MOD_CANONICAL_BASEDIR, matchDll, fileTree);
-
-const red4extBasedirLayout: LayoutToInstructions = (
-  _api: VortexApi,
-  modName: string,
-  fileTree: FileTree,
-): MaybeInstructions => {
-  const hasBasedirFiles = detectRed4ExtBasedirLayout(fileTree);
-
-  if (!hasBasedirFiles) {
-    return NoInstructions.NoMatch;
-  }
-
-  const allFilesUnderBase = findBasedirRed4ExtFiles(fileTree);
-
-  const canonicalModnamedPath = path.join(
-    RED4EXT_MOD_CANONICAL_BASEDIR,
-    modName,
-    path.sep,
-  );
-
-  const allFromBaseToModname: string[][] = allFilesUnderBase.map(
-    moveFromTo(RED4EXT_MOD_CANONICAL_BASEDIR, canonicalModnamedPath),
-  );
-
-  return {
-    kind: Red4ExtLayout.Basedir,
-    instructions: instructionsForSourceToDestPairs(allFromBaseToModname),
-  };
-};
-
-const findCanonicalRed4ExtDirs = (fileTree: FileTree) =>
-  findDirectSubdirsWithSome(RED4EXT_MOD_CANONICAL_BASEDIR, matchDll, fileTree);
-
-const detectRed4ExtCanonOnlyLayout = (fileTree: FileTree): boolean =>
-  !detectRed4ExtBasedirLayout(fileTree) && findCanonicalRed4ExtDirs(fileTree).length > 0;
-
-const red4extCanonLayout: LayoutToInstructions = (
-  api: VortexApi,
-  _modName: string,
-  fileTree: FileTree,
-): MaybeInstructions => {
-  const hasCanonFiles = detectRed4ExtCanonOnlyLayout(fileTree);
-
-  if (!hasCanonFiles) {
-    return NoInstructions.NoMatch;
-  }
-
-  // This is maybe slightly annoying to check, but makes
-  // logic elsewhere cleaner. I suppose we can decide that
-  // layouts need to be robust enough in themselves if they
-  // would otherwise depend on some external check that isn't
-  // always present.
-  //
-  // Generally, we *shouldn't* get here and the problem should
-  // have already been detected in test, but..
-  //
-  const hasBasedirReds = detectRed4ExtBasedirLayout(fileTree);
-
-  if (hasBasedirReds) {
-    // Errors need to be handled downstream if it's relevant there
-    api.log("debug", "No instructions from canon: basedir overrides");
-    return NoInstructions.NoMatch;
-  }
-
-  const allCanonFiles = filesUnder(RED4EXT_MOD_CANONICAL_BASEDIR, Glob.Any, fileTree);
-
-  return {
-    kind: Red4ExtLayout.Canon,
-    instructions: instructionsForSameSourceAndDestPaths(allCanonFiles),
-  };
-};
-
-const red4extToplevelLayout: LayoutToInstructions = (
-  _api: VortexApi,
-  modName: string,
-  fileTree: FileTree,
-): MaybeInstructions => {
-  const toplevelFiles = filesIn(FILETREE_ROOT, matchDll, fileTree);
-
-  if (toplevelFiles.length < 1) {
-    return NoInstructions.NoMatch;
-  }
-
-  // No messing about: a DLL at top level means this entire
-  // thing has to be part of it. Nothing else allowed.
-
-  const allTheFilesEverywhere = filesUnder(FILETREE_ROOT, Glob.Any, fileTree);
-
-  const canonicalModnamedPath = path.join(RED4EXT_MOD_CANONICAL_BASEDIR, modName);
-
-  const allFilesToCanon = allTheFilesEverywhere.map(
-    moveFromTo(FILETREE_ROOT, canonicalModnamedPath),
-  );
-
-  return {
-    kind: Red4ExtLayout.Toplevel,
-    instructions: instructionsForSourceToDestPairs(allFilesToCanon),
-  };
-};
-
-const red4extModnamedToplevelLayout: LayoutToInstructions = (
-  _api: VortexApi,
-  _modName: string,
-  fileTree: FileTree,
-): MaybeInstructions => {
-  const toplevelSubdirsWithFiles = findDirectSubdirsWithSome(
-    FILETREE_ROOT,
-    matchDll,
-    fileTree,
-  );
-
-  if (toplevelSubdirsWithFiles.length < 1) {
-    return NoInstructions.NoMatch;
-  }
-
-  if (toplevelSubdirsWithFiles.length > 1) {
-    return InvalidLayout.Conflict;
-  }
-
-  const allToBasedirWithSubdirAsModname: string[][] = toplevelSubdirsWithFiles.flatMap(
-    (dir) =>
-      filesUnder(dir, Glob.Any, fileTree).map(
-        moveFromTo(FILETREE_ROOT, RED4EXT_MOD_CANONICAL_BASEDIR),
-      ),
-  );
-
-  return {
-    kind: Red4ExtLayout.Modnamed,
-    instructions: instructionsForSourceToDestPairs(allToBasedirWithSubdirAsModname),
-  };
-};
-
-export const testForRed4ExtMod: VortexWrappedTestSupportedFunc = (
-  api: VortexApi,
-  log: VortexLogFunc,
-  files: string[],
-  fileTree: FileTree,
-): Promise<VortexTestResult> => {
-  const allDllSubdirs = findAllSubdirsWithSome(FILETREE_ROOT, matchDll, fileTree);
-  const toplevelDlls = filesIn(FILETREE_ROOT, matchDll, fileTree);
-
-  const noDllDirs = allDllSubdirs.length < 1;
-  const noToplevelDlls = toplevelDlls.length < 1;
-
-  if (noDllDirs && noToplevelDlls) {
-    return Promise.resolve({ supported: false, requiredFiles: [] });
-  }
-
-  const dangerPaths = [
-    ...allDllSubdirs.filter(reservedDllDir),
-    ...toplevelDlls.filter(reservedDllName),
-  ];
-
-  if (dangerPaths.length !== 0) {
-    const message = "Red4Ext Mod Installation Canceled, Dangerous DLL paths!";
-    log("error", message, dangerPaths);
-    showRed4ExtReservedDllErrorDialog(api, message, dangerPaths);
-    return Promise.reject(new Error(message));
-  }
-
-  // Red4Ext itself handled elsewhere
-  if (pathInTree(RED4EXT_CORE_RED4EXT_DLL, fileTree)) {
-    return Promise.resolve({ supported: false, requiredFiles: [] });
-  }
-
-  // Good enough, this is the right installer, more checks in `install`
-
-  return Promise.resolve({ supported: true, requiredFiles: [] });
-};
-
-export const installRed4ExtMod: VortexWrappedInstallFunc = (
-  api: VortexApi,
-  log: VortexLogFunc,
-  files: string[],
-  fileTree: FileTree,
-  destinationPath: string,
-): Promise<VortexInstallResult> => {
-  const modName = makeSyntheticName(destinationPath);
-
-  // At this point we know from the test that none of
-  // these are in dangerous locations
-
-  // Red4Ext only allows one subdir deep for initially loaded DLLs
-  // Move to test?
-
-  // ...And is it a good idea to allow more than one canon subdir?
-  const possibleLayoutsToTryInOrder: LayoutToInstructions[] = [
-    red4extBasedirLayout,
-    red4extCanonLayout,
-    red4extToplevelLayout,
-    red4extModnamedToplevelLayout,
-  ];
-
-  const chosenInstructions = useFirstMatchingLayoutForInstructions(
-    api,
-    modName,
-    fileTree,
-    possibleLayoutsToTryInOrder,
-  );
-
-  if (chosenInstructions === NoInstructions.NoMatch) {
-    const message = "Red4Ext installer failed to generate any instructions!";
-    log("error", message, files);
-    return Promise.reject(new Error(message));
-  }
-
-  if (chosenInstructions === InvalidLayout.Conflict) {
-    return promptToFallbackOrFailOnUnresolvableLayout(
-      api,
-      InstallerType.Red4Ext,
-      fileTree,
-    );
-  }
-
-  const extraArchiveLayoutsAllowed = chosenInstructions.kind !== Red4ExtLayout.Toplevel;
-
-  const allInstructions = extraArchiveLayoutsAllowed
-    ? [
-        ...chosenInstructions.instructions,
-        ...extraCanonArchiveInstructions(api, fileTree).instructions,
-      ]
-    : chosenInstructions.instructions;
-
-  const haveFilesOutsideSelectedInstructions =
-    allInstructions.length !== fileCount(fileTree);
-
-  if (haveFilesOutsideSelectedInstructions) {
-    return promptToFallbackOrFailOnUnresolvableLayout(
-      api,
-      InstallerType.Red4Ext,
-      fileTree,
-    );
-  }
-
-  log("info", "Red4Ext installer installing files.");
-  log("debug", "Red4Ext instructions: ", allInstructions);
-
-  return Promise.resolve({ instructions: allInstructions });
-};
-
-// JSON Mods
-
-export const testForJsonMod: VortexWrappedTestSupportedFunc = (
-  _api: VortexApi,
-  log: VortexLogFunc,
-  files: string[],
-  _fileTree: FileTree,
-): Promise<VortexTestResult> => {
-  const filtered = files.filter(
-    (file: string) => path.extname(file).toLowerCase() === JSON_FILE_EXT,
-  );
-  if (filtered.length === 0) {
-    return Promise.resolve({
-      supported: false,
-      requiredFiles: [],
-    });
-  }
-
-  // This little change should allow properly constructed AMM addons to install in the fallback
-  const cetModJson = files.filter((file: string) =>
-    path.normalize(file).toLowerCase().startsWith(CET_MOD_CANONICAL_PATH_PREFIX),
-  );
-  if (cetModJson.length !== 0) {
-    log("error", "We somehow got a CET mod in the JSON check");
-    return Promise.resolve({
-      supported: false,
-      requiredFiles: [],
-    });
-  }
-
-  // This should probably be moved or made prompting:
-  // https://github.com/E1337Kat/cyberpunk2077_ext_redux/issues/113
-
-  let proper = true;
-  // check for options.json in the file list
-  const options = filtered.some((file: string) => path.basename(file) === "options.json");
-  if (options) {
-    log("debug", "Options.json files found: ", options);
-    proper = filtered.some((f: string) =>
-      path.dirname(f).toLowerCase().startsWith(path.normalize("r6/config/settings")),
-    );
-
-    if (!proper) {
-      const message =
-        "Improperly located options.json file found.  We don't know where it belongs.";
-
-      log("info", message);
-      return Promise.reject(new Error(message));
-    }
-  } else if (
-    filtered.some((file: string) => KNOWN_JSON_FILES[path.basename(file)] === undefined)
-  ) {
-    log("error", "Found JSON files that aren't part of the game.");
-    return Promise.reject(new Error("Found JSON files that aren't part of the game."));
-  }
-
-  log("debug", "We got through it all and it is a JSON mod");
-  return Promise.resolve({
-    supported: true,
-    requiredFiles: [],
-  });
-};
-
-export const installJsonMod: VortexWrappedInstallFunc = (
-  api: VortexApi,
-  log: VortexLogFunc,
-  files: string[],
-  _fileTree: FileTree,
-  _destinationPath: string,
-): Promise<VortexInstallResult> => {
-  const jsonFiles: string[] = files.filter(
-    (file: string) => path.extname(file) === ".json",
-  );
-  const otherAllowedFiles = files.filter(
-    (file: string) => path.extname(file) === ".txt" || path.extname(file) === ".md",
-  );
-
-  const filtered = jsonFiles.concat(otherAllowedFiles);
-
-  let movedJson = false;
-
-  const jsonFileInstructions = filtered.map((file: string) => {
-    const fileName = path.basename(file);
-
-    let instPath = file;
-
-    if (KNOWN_JSON_FILES[fileName] !== undefined) {
-      instPath = KNOWN_JSON_FILES[fileName];
-
-      log("debug", "instPath set as ", instPath);
-      movedJson = movedJson || file !== instPath;
-    }
-
-    return {
-      type: "copy",
-      source: file,
-      destination: instPath,
-    };
-  });
-
-  if (movedJson) {
-    log("info", "JSON files were found outside their canonical locations: Fixed");
-  }
-
-  log("debug", "Installing JSON files with: ", jsonFileInstructions);
-
-  const instructions = [].concat(jsonFileInstructions);
-
-  return Promise.resolve({ instructions });
-};
-
-const testForReshadeFile = (
-  log: VortexLogFunc,
-  files: string[],
-  folder: string,
-): boolean => {
-  // We're going to make a reasonable assumption here that reshades will
-  // only have reshade ini's, so we only need to check the first one
-
-  const fileToExamine = path.join(
-    folder,
-    files.find((file: string) => path.extname(file) === INI_MOD_EXT),
-  );
-
-  const data = fs.readFileSync(fileToExamine, { encoding: "utf8" });
-
-  if (data === undefined) {
-    log("error", "unable to read contents of ", fileToExamine);
-    return false;
-  }
-  data.slice(0, 80);
-  // eslint-disable-next-line no-useless-escape
-  const regex = /^[\[#].+/;
-  const testString = data.replace(regex, "");
-  if (testString === data) {
-    log("info", "Reshade file located.");
-    return true;
-  }
-
-  return false;
-};
-
-// INI (includes Reshade?)
-export const testForIniMod: VortexWrappedTestSupportedFunc = (
-  api: VortexApi,
-  log: VortexLogFunc,
-  files: string[],
-  _fileTree: FileTree,
-): Promise<VortexTestResult> => {
-  const filtered = files.filter(
-    (file: string) => path.extname(file).toLowerCase() === INI_MOD_EXT,
-  );
-
-  if (filtered.length === 0) {
-    log("info", "No INI files.");
-    return Promise.resolve({
-      supported: false,
-      requiredFiles: [],
-    });
-  }
-
-  if (
-    files.some(
-      (file: string) =>
-        path.basename(file).includes(CET_MOD_CANONICAL_INIT_FILE) ||
-        path.extname(file) === REDS_MOD_CANONICAL_EXTENSION,
-    )
-  ) {
-    // These should  actually error out, we should not get here
-    log(`error`, `INI file detected within a CET or Redscript mod, aborting`);
-    return Promise.resolve({
-      supported: false,
-      requiredFiles: [],
-    });
-  }
-  if (files.includes(CET_GLOBAL_INI)) {
-    log(`error`, `CET Installer detected, not processing as INI`);
-    return Promise.resolve({
-      supported: false,
-      requiredFiles: [],
-    });
-  }
-  return Promise.resolve({
-    supported: true,
-    requiredFiles: [],
-  });
-};
-
-export const installIniMod: VortexWrappedInstallFunc = (
-  api: VortexApi,
-  log: VortexLogFunc,
-  files: string[],
-  _fileTree: FileTree,
-  _destinationPath: string,
-): Promise<VortexInstallResult> => {
-  // This installer gets called for both reshade and "normal" ini mods
-
-  const allIniModFiles = files.filter(
-    (file: string) => path.extname(file) === INI_MOD_EXT,
-  );
-
-  const reshade = testForReshadeFile(log, allIniModFiles, _destinationPath);
-
-  // Set destination depending on file type
-
-  log("info", "Installing ini files: ", allIniModFiles);
-  const iniFileInstructions = allIniModFiles.map((file: string) => {
-    const fileName = path.basename(file);
-    const dest = reshade
-      ? path.join(RESHADE_MOD_PATH, path.basename(file))
-      : path.join(INI_MOD_PATH, fileName);
-
-    return {
-      type: "copy",
-      source: file,
-      destination: dest,
-    };
-  });
-
-  const shaderFiles = files.filter(
-    (file: string) => file.includes(RESHADE_SHADERS_DIR) && !file.endsWith(path.sep),
-  );
-
-  let shaderInstructions = [];
-
-  if (reshade && shaderFiles.length !== 0) {
-    log("info", "Installing shader files: ", shaderFiles);
-    shaderInstructions = shaderFiles.map((file: string) => {
-      const regex = /.*reshade-shaders/;
-      const fileName = file.replace(regex, RESHADE_SHADERS_DIR);
-      // log("info", "Shader dir Found. Processing: ", fileName);
-      const dest = path.join(RESHADE_MOD_PATH, fileName);
-      // log("debug", "Shader file: ", dest);
-      return {
-        type: "copy",
-        source: file,
-        destination: dest,
-      };
-    });
-  }
-  const instructions = [].concat(iniFileInstructions, shaderInstructions);
-  log("debug", "Installing ini files with instructions: ", instructions);
-
-  return Promise.resolve({ instructions });
-};
-
-const findCanonicalAsiDirs = (fileTree: FileTree) =>
-  filesIn(ASI_MOD_PATH, matchAsiFile, fileTree);
-
-const detectASICanonLayout = (fileTree: FileTree): boolean =>
-  findCanonicalAsiDirs(fileTree).length > 0;
-
-const findCanonicalAsiFiles = (fileTree: FileTree): string[] =>
-  filesUnder(ASI_MOD_PATH, Glob.Any, fileTree);
-
-export const testForAsiMod: VortexWrappedTestSupportedFunc = (
-  _api: VortexApi,
-  log: VortexLogFunc,
-  files: string[],
-  fileTree: FileTree,
-): Promise<VortexTestResult> => {
-  if (!detectASICanonLayout(fileTree)) {
-    return Promise.resolve({ supported: false, requiredFiles: [] });
-  }
-
-  return Promise.resolve({ supported: true, requiredFiles: [] });
-};
-
-const asiCanonLayout: LayoutToInstructions = (
-  _api: VortexApi,
-  modName: string,
-  fileTree: FileTree,
-): MaybeInstructions => {
-  const hasBasedirFiles = filesIn(ASI_MOD_PATH, matchAsiFile, fileTree).length > 0;
-
-  if (!hasBasedirFiles) {
-    return NoInstructions.NoMatch;
-  }
-
-  const allCanonAsiFiles = findCanonicalAsiFiles(fileTree);
-
-  if (allCanonAsiFiles.length === 0) {
-    return NoInstructions.NoMatch;
-  }
-
-  return {
-    kind: AsiLayout.Canon,
-    instructions: instructionsForSameSourceAndDestPaths(allCanonAsiFiles),
-  };
-};
-
-export const installAsiMod: VortexWrappedInstallFunc = (
-  api: VortexApi,
-  log: VortexLogFunc,
-  files: string[],
-  fileTree: FileTree,
-  destinationPath: string,
-): Promise<VortexInstallResult> => {
-  const modname = makeSyntheticName(destinationPath);
-
-  const chosenInstructions = asiCanonLayout(api, modname, fileTree);
-
-  if (
-    chosenInstructions === NoInstructions.NoMatch ||
-    chosenInstructions === InvalidLayout.Conflict
-  ) {
-    const message = "ASI installer failed to generate instructions";
-    log("error", message, files);
-    return Promise.reject(new Error(message));
-  }
-
-  const { instructions } = chosenInstructions;
-
-  const haveFilesOutsideSelectedInstructions =
-    instructions.length !== fileCount(fileTree);
-
-  if (haveFilesOutsideSelectedInstructions) {
-    const message = `Too many files in ASI Mod! ${instructions.length}`;
-    log("error", message, files);
-    return Promise.reject(new Error(message));
-  }
-
-  log("info", "ASI installer installing files.");
-  log("debug", "ASI instructions: ", instructions);
-
-  return Promise.resolve({ instructions });
-};
-
-// MultiType installer
-
-export const testForMultiTypeMod: VortexWrappedTestSupportedFunc = (
-  api: VortexApi,
-  _log: VortexLogFunc,
-  _files: string[],
-  fileTree: FileTree,
-): Promise<VortexTestResult> => {
-  const hasCanonCet = detectCetCanonLayout(fileTree);
-  const hasCanonRedscript = detectRedscriptCanonOnlyLayout(fileTree);
-  const hasBasedirRedscript = detectRedscriptBasedirLayout(fileTree);
-  const hasCanonRed4Ext = detectRed4ExtCanonOnlyLayout(fileTree);
-  const hasBasedirRed4Ext = detectRed4ExtBasedirLayout(fileTree);
-
-  // The Onlys may need better naming.. they already check that
-  // there's no basedir stuff, so we can use both here without
-  // additional checks.
-  const hasAtLeastTwoTypes =
-    [
-      hasCanonCet,
-      hasCanonRedscript,
-      hasBasedirRedscript,
-      hasCanonRed4Ext,
-      hasBasedirRed4Ext,
-    ].filter(trueish).length > 1;
-
-  if (!hasAtLeastTwoTypes) {
-    api.log("debug", "MultiType didn't match");
-    return Promise.resolve({ supported: false, requiredFiles: [] });
-  }
-
-  api.log("info", "MultiType mod detected", {
-    hasCanonCet,
-    hasCanonRedscript,
-    hasBasedirRedscript,
-    hasCanonRed4Ext,
-    hasBasedirRed4Ext,
-  });
-
-  return Promise.resolve({
-    supported: true,
-    requiredFiles: [],
-  });
-};
-
-export const installMultiTypeMod: VortexWrappedInstallFunc = (
-  api: VortexApi,
-  _log: VortexLogFunc,
-  _files: string[],
-  fileTree: FileTree,
-  destinationPath: string,
-): Promise<VortexInstallResult> => {
-  // Should extract this to wrapper..
-  const modName = makeSyntheticName(destinationPath);
-
-  // This should be more robust. Currently we kinda rely
-  // on it being very unlikely that these kinds of mods
-  // are broken in ways like having canon and basedir
-  // stuff but that's not guaranteed.
-  //
-  // Defect: https://github.com/E1337Kat/cyberpunk2077_ext_redux/issues/96
-
-  // Also notable: Basedirs currently *override* Canon.
-  // This is probably the desired behavior but I dunno
-  // if we could at least make it somehow more obvious
-  // in the naming scheme.. it's clearer in specific
-  // installers where we choose one layout only.
-  const allInstructionSets: LayoutToInstructions[] = [
-    cetCanonLayout,
-    redscriptBasedirLayout,
-    redscriptCanonLayout,
-    red4extBasedirLayout,
-    red4extCanonLayout,
-    archiveCanonLayout,
-    archiveHeritageLayout,
-  ];
-
-  const allInstructionsPerLayout = useAllMatchingLayouts(
-    api,
-    modName,
-    fileTree,
-    allInstructionSets,
-  );
-
-  const allInstructions = allInstructionsPerLayout.flatMap((i) => i.instructions);
-
-  // Should still add this..
-  // warnUserIfArchivesMightNeedManualReview(api, instrs stils needed);
-  const haveFilesOutsideSelectedInstructions =
-    allInstructions.length !== fileCount(fileTree);
-
-  if (allInstructionsPerLayout.length < 1 || haveFilesOutsideSelectedInstructions) {
-    return promptToFallbackOrFailOnUnresolvableLayout(
-      api,
-      InstallerType.MultiType,
-      fileTree,
-    );
-  }
-
-  api.log("info", "MultiType installer installing files.");
-  api.log("debug", "MultiType instructions: ", allInstructions);
-
-  return Promise.resolve({ instructions: allInstructions });
-};
 
 // Setup stuff, pipeline
 
@@ -1213,15 +160,6 @@ const installers: Installer[] = [
     testSupported: testForMultiTypeMod,
     install: installMultiTypeMod,
   },
-  // Remove this once the mixes are fully moved,
-  // currently this CET logic actually
-  // Deprecate: https://github.com/E1337Kat/cyberpunk2077_ext_redux/issues/82
-  // {
-  //   type: InstallerType.RedCetMix,
-  //   id: "cp2077-red-cet-mixture-mod",
-  //   testSupported: testForRedCetMixedMod,
-  //   install: installRedCetMixedMod,
-  // },
   {
     type: InstallerType.CET,
     id: InstallerType.CET,
@@ -1264,12 +202,6 @@ const installers: Installer[] = [
   {
     type: InstallerType.Config,
     id: "cp2077-config-mod",
-    testSupported: notSupportedModType,
-    install: notInstallableMod,
-  },
-  {
-    type: InstallerType.Reshade,
-    id: "cp2077-reshade-mod",
     testSupported: notSupportedModType,
     install: notInstallableMod,
   },
